chore(replay): move wrappers out of init
